refactor: separate bussiness login from service

+++ Backend/src/todo/todo.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateTodoDto } from './dto/create-todo.dto';
 import { UpdateTodoDto } from './dto/update-todo.dto';
 import { Item } from '@prisma/client';
@@ -9,6 +9,10 @@ export class TodoService {
   constructor(private readonly baseService: BaseService<Item>) {}
 
   create(createTodoDto: CreateTodoDto): Promise<Item> {
+
+    if (createTodoDto.dueDate)
+      createTodoDto.dueDate = new Date(createTodoDto.dueDate);
+
     return this.baseService.create(createTodoDto);
   }
 
@@ -21,10 +25,18 @@ export class TodoService {
   }
 
   update(id: string, updateTodoDto: UpdateTodoDto): Promise<Item> {
+
+    if (updateTodoDto.dueDate)
+      updateTodoDto.dueDate = new Date(updateTodoDto.dueDate);
+
     return this.baseService.update(id, updateTodoDto);
   }
 
-  remove(id: string): Promise<Item> {
+  async remove(id: string): Promise<Item> {
+
+    const exitingItem = await this.baseService.findOne(id);
+    if (!exitingItem) throw new NotFoundException('Item not found');
+
     return this.baseService.remove(id);
   }
 }
